+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Not released
 
 - Improve Widgets calculations sync with tiles [#223](https://github.com/CartoDB/carto-react/pull/223)
+- `CategoryWidget`, `HistogramWidget`, `PieWidget` use filters in datasource to derive selecteditems (categories, bars) instead of local react state [224](https://github.com/CartoDB/carto-react/pull/224)
 
 ## 1.1.0 (2021-10-29)